rename cmd libraries and subdirectories as cli
