+++ cmd/dev/backend_kv_server.cpp
@@ -33,10 +33,10 @@
 #include <silkworm/core/types/evmc_bytes32.hpp>
 #include <silkworm/db/access_layer.hpp>
 #include <silkworm/db/chain_head.hpp>
-#include <silkworm/db/cmd/db_max_readers_option.hpp>
+#include <silkworm/db/cli/db_max_readers_option.hpp>
 #include <silkworm/db/datastore/kvdb/mdbx.hpp>
-#include <silkworm/infra/cmd/common.hpp>
-#include <silkworm/infra/cmd/shutdown_signal.hpp>
+#include <silkworm/infra/cli/common.hpp>
+#include <silkworm/infra/cli/shutdown_signal.hpp>
 #include <silkworm/infra/common/directories.hpp>
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/infra/concurrency/awaitable_wait_for_one.hpp>
