+++ silkworm/db/cli/CMakeLists.txt
@@ -19,12 +19,7 @@ include("${SILKWORM_MAIN_DIR}/cmake/common/targets.cmake")
 find_package(Boost REQUIRED COMPONENTS headers)
 find_package(CLI11 REQUIRED)
 
-set(LIBS_PUBLIC
-    silkworm_core
-    silkworm_infra
-    Boost::headers
-    CLI11::CLI11
-)
+set(LIBS_PUBLIC silkworm_db silkworm_infra_cli Boost::headers CLI11::CLI11)
 
 # cmake-format: off
 set(LIBS_PRIVATE
@@ -33,7 +28,7 @@ set(LIBS_PRIVATE
 # cmake-format: on
 
 silkworm_library(
-    silkworm_infra_cmd
-    PUBLIC ${LIBS_PUBLIC}
-    PRIVATE ${LIBS_PRIVATE}
+  silkworm_db_cli
+  PUBLIC ${LIBS_PUBLIC}
+  PRIVATE ${LIBS_PRIVATE}
 )
