+++ silkworm/node/cli/CMakeLists.txt
@@ -16,24 +16,8 @@
 
 include("${SILKWORM_MAIN_DIR}/cmake/common/targets.cmake")
 
-find_package(Boost REQUIRED COMPONENTS headers)
 find_package(CLI11 REQUIRED)
 
-set(LIBS_PUBLIC
-    silkworm_db
-    silkworm_infra_cmd
-    Boost::headers
-    CLI11::CLI11
-)
+set(LIBS_PUBLIC silkworm_core silkworm_node silkworm_infra_cli CLI11::CLI11)
 
-# cmake-format: off
-set(LIBS_PRIVATE
-    silkworm-buildinfo
-)
-# cmake-format: on
-
-silkworm_library(
-    silkworm_db_cmd
-    PUBLIC ${LIBS_PUBLIC}
-    PRIVATE ${LIBS_PRIVATE}
-)
+silkworm_library(silkworm_node_cli PUBLIC ${LIBS_PUBLIC})
