+++ silkworm/sentry/cli/CMakeLists.txt
@@ -14,7 +14,7 @@
    limitations under the License.
 ]]
 
-set(TARGET silkworm_sentry_cmd)
+set(TARGET silkworm_sentry_cli)
 
 find_package(CLI11 REQUIRED)
 
@@ -22,4 +22,4 @@ file(GLOB_RECURSE SRC CONFIGURE_DEPENDS "*.cpp" "*.hpp")
 
 add_library(${TARGET} ${SRC})
 
-target_link_libraries(${TARGET} PUBLIC silkworm_core silkworm_infra silkworm_sentry silkworm_infra_cmd CLI11::CLI11)
+target_link_libraries(${TARGET} PUBLIC silkworm_core silkworm_infra silkworm_sentry silkworm_infra_cli CLI11::CLI11)
